Support Maps place marker on selected destination
1. Feature: on clicking listview item, open map with a placemark on the
destination selected.
2. Implemented destination object as Serializable because we need to
pass destination information from destination list to map activity. For
inter-activity communication, an object should be either Serializable or
parceable.
3. Bug Fix: Latitude and Longitude were retrieved as integer instead of
double because of that getting 78.00 instead of exact location
78.98593453.

+++ src/com/csc591/DAL/DestinationDataSource.java
@@ -109,8 +109,13 @@ public class DestinationDataSource {
 	{
 		Destination dest = new Destination();
 		dest.setId(cursor.getLong(0));
-		dest.setLatitude(cursor.getInt(1));
-		dest.setLongitude(cursor.getInt(2));
+		// BUG FIX: In SQLite everything is stored as text however you need to take extra care
+		// while fetching data from the database. Although while storing the Latitude, Longitude 
+		// data type is INTEGER (see MySQLiteHelper) but actually Lat, Lng are double values. 
+		// while fetching data, use proper data type.
+		// Credits: http://stackoverflow.com/questions/3083796/storing-double-values-in-sqlite-how-to-ensure-precision?answertab=votes#tab-top
+		dest.setLatitude(cursor.getDouble(1));
+		dest.setLongitude(cursor.getDouble(2));
 		dest.setName(cursor.getString(3));
 		dest.setType(cursor.getInt(4));
 		dest.setDescription(cursor.getString(5));
